backend_demo/requirements.txt
annotated-types==0.7.0
asgiref==3.8.1
black==24.4.2
click==8.1.7
contextlib2==21.6.0
Django==5.0.6
django-extensions==3.2.3
django-ninja==1.2.0
django-ninja-extra==0.21.0
injector==0.21.0
mypy-extensions==1.0.0
packaging==24.1
pathspec==0.12.1
platformdirs==4.2.2
pydantic==2.8.2
pydantic_core==2.20.1
sqlparse==0.5.0
typing_extensions==4.12.2
